+++ courses/prog_base/labs/lab2/loops.c
@@ -18,7 +18,7 @@ double calc(int n, int m){
     int i, j;
     for(i = 1; i <= n; i++){
         for(j = 1; j <= m; j++){
-            x = x + cos(M_PI/i)*(i/j);
+            x = x + cos(M_PI/i)*((double)i/j);
         }
     }
     return x;
